Guards against setValue forKey crashes.
When an object changes schema between having been serialized (on app update for instance), if we've added _Nonnull properties, when we use Swift, we end up with a Fatal Exception: NSInvalidArgumentException [ setNilValueForKey]: could not set nil as the value for the key ..." that cannot be caught at Swift level.
Wrapping the setValue forKey in @try @catch  allows catching the error without breaking the app.